refactor: 리팩토링한 위치에 맞춰 test코드 리팩토링
